# application/services/custom_services/gepphysicaladdressmanager/src/models/entitymodels/LocationTags.ts
export interface LocationTags 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   latitude: String,
   longitude: String,
   geoname: String
}