# application/services/custom_services/gepphysicaladdressmanager/src/swagger/swagger.yaml
---
openapi: 3.0.0
info:
   version: 1.3.0
   title: angulartest open api specification
   description: gepphysicaladdressmanager API
   termsOfService: http://api_url/terms/
   contact:
      name: ContactName
      email: hello@wolox.co
      url: https://www.wolox.com.ar/
   license:
      name: Apache 2.0
      url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
   - url: http://localhost:0
     description: gepphysicaladdressmanager
tags:
- name: LocationTags
  description: Description
paths:
   "/LocationTags/{id}":
      delete:
         tags:
         -  LocationTags
         description: GpDelete
         parameters:
         -  in: "path"
            description: GpDelete
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      get:
         tags:
         -  LocationTags
         description: GpGetNounById
         parameters:
         -  in: "path"
            description: GpGetNounById
            name: id
            required: true
            schema:
               type: "string"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/LocationTags":
      get:
         tags:
         -  LocationTags
         description: GpGetAllValues
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      post:
         tags:
         -  LocationTags
         description: GpCreate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/LocationTags"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
      put:
         tags:
         -  LocationTags
         description: GpUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/LocationTags"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/LocationTags/get/search":
      get:
         tags:
         -  LocationTags
         description: GpSearch
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/LocationTags/get/update":
      put:
         tags:
         -  LocationTags
         description: GpSearchForUpdate
         requestBody:
            content:
               application/json:
                  schema:
                     $ref: "#/components/schemas/LocationTags"
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/LocationTags/get/searchrelationship":
      get:
         tags:
         -  LocationTags
         description: GpSearchRelationship
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
   "/LocationTags/userid/created_by":
      get:
         tags:
         -  LocationTags
         description: GpGetNounCreatedBy
         responses:
            200:
               description: Success
            400:
               description: Bad Request
            404:
               description: Not Found
            500:
               description: Internal Server Error
components:
   schemas:
      LocationTags:
         type: object
         properties:
            latitude:
               type: "string"
               description: Description
            longitude:
               type: "string"
               description: Description
            geoname:
               type: "string"
               description: Description